+++ CHANGELOG.rst
@@ -50,6 +50,10 @@ This version will be backwards incompatible, and only support Python 3.6+.
 
 - Other changes:
 
+  - ``bipole`` and ``dipole`` now only have the following mandatory positional
+    arguments: ``src``, ``rec``, ``depth``, ``res``, and ``freqtime``, and the
+    optional positional keyword arguments ``signal`` and ``aniso``; all the
+    others are now within ``**kwargs``.
   - ``htarg`` and ``ftarg`` internally are now dictionaries, not lists.
   - Undo a change introduced in v1.8.0: ``get_dlf_points`` is calculated
     directly within ``transform.fht`` [`empymod#26